move allocator struct to new allocator.rs
